@@ -50,8 +50,8 @@ class _MiddleCartState extends State<MiddleCart> {
   }
 
   productscreen(List data) {
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      return Navigator.pushReplacement(
+    WidgetsBinding.instance?.addPostFrameCallback((_) {
+       Navigator.pushReplacement(
           context,
           MaterialPageRoute(
               builder: (context) => PurchaseScreen(
